Split environment-window.tsx into feature-scoped components
The 2015-line file (the "Containers" window's entry point) mixed 8
config tabs, a status card, delete confirmation, native-mode logs
streaming, and shared form primitives in one component. Split into
src/features/environment-window/ following the same shape as the
earlier containers.rs decomposition: each tab becomes its own file
sharing the orchestrator's draft/update(Environment) contract, dialog
state moves to whichever tab owns it, and busy/message/inspect() stay
in the orchestrator since they're genuinely cross-cutting.

No behavior change intended - same props, same Tauri commands, same
emit/listen wiring (environments-changed after save/remove/operate,
environment-progress listener, environment-log-line streaming).
