+++ src/features/environment-window/constants.ts
@@ -1,0 +1,88 @@
+import type { Environment } from "@/types"
+
+export const defaults = (): Environment => ({
+  id: `env-${Date.now()}`,
+  name: "local",
+  webServer: "Nginx",
+  webVersion: "1.28",
+  phpVersion: "8.4",
+  database: "MariaDB",
+  databaseVersion: "11.8",
+  port: "8080",
+  webContainerName: "",
+  databaseContainerName: "",
+  phpExtensions: ["intl", "mbstring", "opcache", "zip"],
+  extraServices: [],
+  nodeVersion: "22",
+  redisVersion: "7.4",
+  databaseName: "app",
+  databaseUser: "app",
+  databasePassword: "localdev",
+  databaseRootPassword: "localdev",
+  phpMemoryLimit: "256M",
+  phpUploadLimit: "64M",
+  phpPostLimit: "64M",
+  phpExecutionTime: 120,
+  phpJit: false,
+  phpJitMode: "tracing",
+  phpJitBufferSize: "64M",
+  phpCron: false,
+  phpCronSchedule: "* * * * *",
+  phpCronCommand: "php artisan schedule:run",
+  phpFpmProcessManager: "dynamic",
+  phpFpmMaxChildren: 10,
+  phpFpmStartServers: 2,
+  phpFpmMinSpareServers: 1,
+  phpFpmMaxSpareServers: 3,
+  phpFpmMaxRequests: 500,
+  phpXdebug: false,
+  phpXdebugMode: "develop,debug",
+  phpXdebugPort: 9003,
+  phpXdebugStart: "trigger",
+  phpXdebugIdeKey: "LSPANEL",
+  environmentVariables: {},
+  redisPassword: "",
+  redisMemoryLimit: "256mb",
+  redisEvictionPolicy: "allkeys-lru",
+  elasticsearchVersion: "8.15.3",
+  elasticsearchMemoryLimit: "512m",
+  minioVersion: "RELEASE.2024-11-07T00-52-20Z",
+  minioRootUser: "minioadmin",
+  minioRootPassword: "minioadmin123",
+  rabbitmqVersion: "3.13",
+  rabbitmqUser: "guest",
+  rabbitmqPassword: "guest",
+  nodePackageManager: "npm",
+  nodeAutoInstall: true,
+  nodeAutoRestart: true,
+  nodeCommand: "",
+  nodeRunMode: "dev",
+  nodeDevCommand: "npm run dev -- --host 0.0.0.0",
+  nodeBuildCommand: "npm run build",
+  nodeStartCommand: "npm start",
+  nodeInspector: false,
+  nodeInspectorPort: 9229,
+  composerVersion: "2",
+  restartPolicy: "unless-stopped",
+  cpuLimit: "2.0",
+  containerMemoryLimit: "2g",
+  runtimeMode: "container",
+  backupScheduleEnabled: false,
+  backupScheduleIntervalHours: 24,
+  backupRetentionCount: 7,
+  wordpressSiteTitle: "",
+  wordpressAdminUser: "",
+  wordpressAdminPassword: "",
+  wordpressAdminEmail: "",
+})
+
+export const services = [
+  "redis",
+  "node",
+  "mailpit",
+  "adminer",
+  "phpmyadmin",
+  "elasticsearch",
+  "minio",
+  "rabbitmq",
+]
