+++ src/features/environment-window/form-fields.tsx
@@ -1,0 +1,159 @@
+import * as React from "react"
+import { Check, Copy, Eye, EyeOff } from "lucide-react"
+
+import { Button } from "@/components/ui/button"
+import { Checkbox } from "@/components/ui/checkbox"
+import { Input } from "@/components/ui/input"
+import { Label } from "@/components/ui/label"
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select"
+import { pickLanguage } from "@/i18n"
+
+export function Field({ label, children }: { label: string; children: React.ReactNode }) {
+  return (
+    <div className="grid gap-2">
+      <Label>{label}</Label>
+      {children}
+    </div>
+  )
+}
+
+export function SecretInput({
+  value,
+  language,
+  onChange,
+}: {
+  value: string
+  language: string
+  onChange: (value: string) => void
+}) {
+  const text = pickLanguage(language).environmentWindow
+  const [visible, setVisible] = React.useState(false)
+  const [copied, setCopied] = React.useState(false)
+  return (
+    <div className="flex gap-2">
+      <Input
+        type={visible ? "text" : "password"}
+        value={value}
+        onChange={(event) => onChange(event.target.value)}
+      />
+      <Button
+        type="button"
+        variant="outline"
+        size="icon"
+        title={visible ? text.hidePassword : text.showPassword}
+        onClick={() => setVisible((current) => !current)}
+      >
+        {visible ? <EyeOff /> : <Eye />}
+      </Button>
+      <Button
+        type="button"
+        variant="outline"
+        size="icon"
+        title={copied ? text.copied : text.copyPassword}
+        disabled={!value}
+        onClick={() => {
+          void navigator.clipboard.writeText(value).then(() => {
+            setCopied(true)
+            window.setTimeout(() => setCopied(false), 1500)
+          })
+        }}
+      >
+        {copied ? <Check /> : <Copy />}
+      </Button>
+    </div>
+  )
+}
+
+export function Choice({
+  value,
+  values,
+  onChange,
+}: {
+  value: string
+  values: string[]
+  onChange: (value: string) => void
+}) {
+  return (
+    <Select value={value} onValueChange={(next) => next && onChange(String(next))}>
+      <SelectTrigger className="w-full">
+        <SelectValue />
+      </SelectTrigger>
+      <SelectContent>
+        {values.map((item) => (
+          <SelectItem key={item} value={item}>
+            {item}
+          </SelectItem>
+        ))}
+      </SelectContent>
+    </Select>
+  )
+}
+export function NumberField({
+  label,
+  value,
+  onChange,
+}: {
+  label: string
+  value: number
+  onChange: (value: number) => void
+}) {
+  return (
+    <Field label={label}>
+      <Input
+        type="number"
+        min={0}
+        value={value}
+        onChange={(event) => onChange(Number(event.target.value))}
+      />
+    </Field>
+  )
+}
+export function CheckRow({
+  label,
+  checked,
+  onChange,
+}: {
+  label: string
+  checked: boolean
+  onChange: (value: boolean) => void
+}) {
+  return (
+    <label className="flex items-center gap-2 text-sm">
+      <Checkbox checked={checked} onCheckedChange={(value) => onChange(Boolean(value))} />
+      {label}
+    </label>
+  )
+}
+export function CheckGrid({
+  title,
+  values,
+  selected,
+  onToggle,
+}: {
+  title: string
+  values: string[]
+  selected: string[]
+  onToggle: (value: string) => void
+}) {
+  return (
+    <div className="grid gap-3">
+      <Label>{title}</Label>
+      <div className="grid gap-2 sm:grid-cols-3">
+        {values.map((value) => (
+          <CheckRow
+            key={value}
+            label={value}
+            checked={selected.includes(value)}
+            onChange={() => onToggle(value)}
+          />
+        ))}
+      </div>
+    </div>
+  )
+}
